Edit delete movie handler logic

+++ movies-now/src/handlers/deleteMovieHandler.js
@@ -1,7 +1,11 @@
+import { DELETE_MOVIE_CONFIRM_MESSAGE } from '../messages/confirmMessage';
 import { deleteMovie } from '../services/movieService';
 
-export const deleteMovieHandler = () => {
-    deleteMovie(movieCardId);
-    console.log('Hi');
-    // navigate('/movies');
+export const deleteMovieHandler = (movieCardId, navigate) => {
+    const confirmed = window.confirm(DELETE_MOVIE_CONFIRM_MESSAGE);
+
+    if (confirmed) {
+        deleteMovie(movieCardId);
+        navigate('/movies');
+    }
 }
